+++ CHANGELOG.md
@@ -1,16 +1,16 @@
 # Changelog
 
-## [1.2.0] 2024-05-03
+## [1.2.1] 2024-05-03
 
 - Added new endpoints with examples including resize, verify, webhooks, and more. Please see ReadMe for details.
 - Added new test cases.
 - Switch endpoint calls from app.ayrshare.com to api.ayrshare.com.
-- Moved code to support ES Module.
+- Moved code to support ES Module, but also still CommonJS.
 - Updated packages.
 
 ### Breaking Changes
 
-Renamed the package from `social-post-api` to `social-media-api`. Be sure to update your package require/import. 
+Renamed the package from `social-post-api` to `social-media-api`. Be sure to update your package require/import.
 
 Why the change? Because as Ayrshare has matured we've gone well beyond just posting.